set the 1st array length to fit the treeType array length

+++ src/edu/unca/csci201/Tree.java
@@ -31,7 +31,7 @@ public class Tree {
 		return height;
 	}
 
-	public void setHeight(int height) {
+	public void setHeight(double height) {
 		this.height = height;
 	}
 
@@ -39,7 +39,7 @@ public class Tree {
 		return width;
 	}
 
-	public void setWidth(int width) {
+	public void setWidth(double width) {
 		this.width = width;
 	}
